Revert "Refactor: moved business logic to Use Cases and added unit test suite"
