+++ app/src/main/java/com/pranshulgg/weather_master_app/feature/main/MainScreen.kt
@@ -167,6 +167,7 @@ fun MainScreen(navController: NavController, weatherViewModel: WeatherViewModel)
                     if (activeLocation != null) {
                         weatherViewModel.getWeather(
                             activeLocation,
+                            activeLocation.source,
                             isManualRefresh = true
                         )
                     }
@@ -200,3 +201,6 @@ fun MainScreen(navController: NavController, weatherViewModel: WeatherViewModel)
         show = mainScreenUiState.isChangelogSheetOpen
     )
 }
+
+
+
