+++ app/src/main/java/com/pranshulgg/weather_master_app/feature/shared/WeatherViewModel.kt
@@ -22,13 +22,11 @@ import com.pranshulgg.weather_master_app.core.model.weather.alerts.AlertResult
 import com.pranshulgg.weather_master_app.core.model.weather.openmeteo.OpenMeteoModel
 import com.pranshulgg.weather_master_app.core.ui.snackbar.SnackbarManager
 import com.pranshulgg.weather_master_app.data.repository.LocationsRepository
+import com.pranshulgg.weather_master_app.data.repository.WeatherBlocksRepository
+import com.pranshulgg.weather_master_app.data.repository.WeatherDataReconcilerRepository
 import com.pranshulgg.weather_master_app.data.repository.WeatherUnitsRepository
+import com.pranshulgg.weather_master_app.data.repository.data.SourceDataRepository
 import com.pranshulgg.weather_master_app.data.worker.WeatherBackgroundUpdateScheduler
-import com.pranshulgg.weather_master_app.domain.usecase.DeleteLocationUseCase
-import com.pranshulgg.weather_master_app.domain.usecase.GetWeatherUseCase
-import com.pranshulgg.weather_master_app.domain.usecase.LoadWeatherBlocksUseCase
-import com.pranshulgg.weather_master_app.domain.usecase.SaveWeatherBlocksUseCase
-import com.pranshulgg.weather_master_app.domain.usecase.UpdateLocationSourceUseCase
 import com.pranshulgg.weather_master_app.feature.main.MainScreenWeatherUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dagger.hilt.android.qualifiers.ApplicationContext
@@ -48,20 +46,14 @@ import kotlin.time.Duration.Companion.minutes
 class WeatherViewModel @Inject constructor(
     private val locationsRepo: LocationsRepository,
     appWeatherUnitsRepo: WeatherUnitsRepository,
-    private val getWeatherUseCase: GetWeatherUseCase,
-    private val deleteLocationUseCase: DeleteLocationUseCase,
-    private val updateLocationSourceUseCase: UpdateLocationSourceUseCase,
-    private val loadWeatherBlocksUseCase: LoadWeatherBlocksUseCase,
-    private val saveWeatherBlocksUseCase: SaveWeatherBlocksUseCase,
+    private val weatherBlocksRepository: WeatherBlocksRepository,
+    private val weatherDataReconcilerRepository: WeatherDataReconcilerRepository,
+    private val sourceDataRepository: SourceDataRepository,
     @ApplicationContext private val context: Context
 ) : ViewModel() {
 
     private var _uiState = mutableStateOf(MainScreenWeatherUiState())
     val uiState: State<MainScreenWeatherUiState> = _uiState
-
-    companion object {
-        private val AUTO_REFRESH_INTERVAL = 45.minutes
-    }
 
     // Registered on the process-wide lifecycle (same pattern as AppVisibility) rather than a
     // Compose LocalLifecycleOwner tied to a screen: a screen-scoped observer gets torn down and
@@ -71,11 +63,11 @@ class WeatherViewModel @Inject constructor(
     // this ViewModel's lifetime, so this observer is only ever added once.
     private val processLifecycleObserver = object : DefaultLifecycleObserver {
         override fun onStart(owner: LifecycleOwner) {
-            startAutoRefresh()
             val location = _uiState.value.activeLocation ?: return
+            startAutoRefresh(location = location, source = location.source)
             // isInitialized guard avoids duplicating setActiveLocation()'s cold-start fetch.
             if (_uiState.value.isInitialized) {
-                getWeather(location = location)
+                getWeather(location = location, source = location.source)
             }
         }
 
@@ -138,6 +130,7 @@ class WeatherViewModel @Inject constructor(
 
     fun getWeather(
         location: Location,
+        source: Source,
         isManualRefresh: Boolean = false,
         isForceRefresh: Boolean = false,
         isForceRefreshForAirQuality: Boolean = false,
@@ -155,16 +148,35 @@ class WeatherViewModel @Inject constructor(
 
         weatherJob = viewModelScope.launch {
 
-            getWeatherUseCase(
-                location = location,
+            var effectiveLocation = location
+            var effectiveForceRefresh = isForceRefresh
+            var effectiveForceRefreshForAirQuality = isForceRefreshForAirQuality
+            var effectiveForceRefreshForAlerts = isForceRefreshForAlerts
+
+            // Checked regardless of isManualRefresh so it also runs on app-open/auto-refresh,
+            // not just pull-to-refresh. If the device actually moved, force a real fetch for
+            // the new coordinates instead of trusting a cache keyed to the old ones (note:
+            // isForceRefresh bypasses the cache unconditionally, unlike isManualRefresh, which
+            // only relaxes the cache TTL and would still be blocked by the 15-min throttle).
+            if (location.isDeviceLocation) {
+                val positionChanged = handleDeviceLocation()
+                if (positionChanged) {
+                    effectiveLocation = locationsRepo.getLocationForId(location.id)
+                    effectiveForceRefresh = true
+                    effectiveForceRefreshForAirQuality = true
+                    effectiveForceRefreshForAlerts = true
+                    _uiState.value = _uiState.value.copy(activeLocation = effectiveLocation)
+                }
+            }
+
+
+            sourceDataRepository.getData(
+                location = effectiveLocation,
                 isManualRefresh = isManualRefresh,
-                isForceRefresh = isForceRefresh,
-                isForceRefreshForAirQuality = isForceRefreshForAirQuality,
-                isForceRefreshForAlerts = isForceRefreshForAlerts,
-                onLocationUpdated = { updatedLocation ->
-                    _uiState.value = _uiState.value.copy(activeLocation = updatedLocation)
-                },
-                onWeather = { result, effectiveLocation ->
+                isForceRefresh = effectiveForceRefresh,
+                isForceRefreshForAirQuality = effectiveForceRefreshForAirQuality,
+                isForceRefreshForAlerts = effectiveForceRefreshForAlerts,
+                onWeather = { result ->
                     handleWeatherData(result, effectiveLocation)
                 },
                 onAlerts = { result ->
@@ -174,6 +186,34 @@ class WeatherViewModel @Inject constructor(
                     handleAirQuality(result)
                 },
             )
+//            handleWeatherData(source, effectiveLocation, isManualRefresh, effectiveForceRefresh)
+
+            // Run separately
+//            if (!_uiState.value.isError) {
+//                launch {
+//                    handleAirQuality(
+//                        effectiveLocation,
+//                        isManualRefresh,
+//                        effectiveForceRefreshForAirQuality
+//                    )
+//                }
+//
+
+            // Only run separately if the alert source is a different API of its own
+            // If not, then run after the weather has fetched
+//                if (!source.providesAlerts && effectiveLocation.alertSource.name != source.name) {
+//                    launch {
+//                        handleAlerts(
+//                            effectiveLocation,
+//                            isManualRefresh,
+//                            effectiveForceRefreshForAlerts
+//                        )
+//                    }
+//                } else {
+//                    handleAlertsFromWeatherSource(effectiveLocation)
+//                }
+//            }
+
 
             val elapsed = System.currentTimeMillis() - startTime
             val minLoadingTime = 1000L // 1s
@@ -194,7 +234,7 @@ class WeatherViewModel @Inject constructor(
 
     fun deleteLocation(id: String) {
         viewModelScope.launch {
-            deleteLocationUseCase(id)
+            locationsRepo.deleteLocation(id)
 
             if (_uiState.value.activeLocation?.id == id) {
                 setActiveLocation(_uiState.value.locations.first { it.isDefault })
@@ -209,7 +249,7 @@ class WeatherViewModel @Inject constructor(
 
     fun setActiveLocation(location: Location) {
         _uiState.value = _uiState.value.copy(activeLocation = location)
-        getWeather(location)
+        getWeather(location, location.source)
     }
 
 
@@ -220,24 +260,33 @@ class WeatherViewModel @Inject constructor(
         alertSource: Source,
         openMeteoModel: OpenMeteoModel
     ) {
+        val updatedLocation = location.copy(
+            source = source,
+            airQualitySource = airQualitySource,
+            alertSource = alertSource,
+            openMeteoModel = openMeteoModel
+        )
+
         viewModelScope.launch {
 
-            val updatedLocation = updateLocationSourceUseCase(
-                location = location,
-                source = source,
-                airQualitySource = airQualitySource,
-                alertSource = alertSource,
-                openMeteoModel = openMeteoModel
-            )
+            locationsRepo.updateSourceForLocation(location.id, source)
+            locationsRepo.updateAirQualitySourceForLocation(location.id, airQualitySource)
+            locationsRepo.updateAlertSourceForLocation(location.id, alertSource)
+            locationsRepo.updateOpenMeteoModelForLocation(location.id, openMeteoModel)
 
             val allowForceRefreshForWeather =
                 location.source != source || location.openMeteoModel != openMeteoModel
             val allowForceRefreshForAirQuality = location.airQualitySource != airQualitySource
             val allowForceRefreshForAlerts = location.alertSource != alertSource
 
+            weatherDataReconcilerRepository.reconcileSourceChange(
+                previous = location,
+                updated = updatedLocation
+            )
             _uiState.value = _uiState.value.copy(activeLocation = updatedLocation)
             getWeather(
                 updatedLocation,
+                source,
                 isForceRefresh = allowForceRefreshForWeather,
                 isForceRefreshForAirQuality = allowForceRefreshForAirQuality,
                 isForceRefreshForAlerts = allowForceRefreshForAlerts
@@ -252,7 +301,15 @@ class WeatherViewModel @Inject constructor(
     ) {
 
         viewModelScope.launch {
-            saveWeatherBlocksUseCase(items, isDaily)
+            weatherBlocksRepository.saveBlocks(items.map {
+                WeatherBlock(
+                    type = it.type,
+                    isHidden = false,
+                    position = it.position,
+                    isDaily = isDaily,
+                    id = it.id
+                )
+            }, isDaily)
 
         }
         _uiState.value = _uiState.value.copy(blocks = items)
@@ -260,8 +317,13 @@ class WeatherViewModel @Inject constructor(
     }
 
     suspend fun loadBlocks() {
-        val loadedBlocks = loadWeatherBlocksUseCase()
+        val loadedBlocks = weatherBlocksRepository.loadBlocks()
         _uiState.value = _uiState.value.copy(blocks = loadedBlocks)
+    }
+
+
+    private suspend fun handleDeviceLocation(): Boolean {
+        return locationsRepo.updateDeviceLocationPosition()
     }
 
     private suspend fun handleWeatherData(result: WeatherResult, location: Location) {
@@ -275,8 +337,6 @@ class WeatherViewModel @Inject constructor(
 
             is WeatherResult.Error -> {
 
-                val appExpectation = result.exception.toAppException()
-                SnackbarManager.show(appExpectation.toMessageRes())
 
                 _uiState.value = _uiState.value.copy(
                     isError = true,
@@ -285,6 +345,9 @@ class WeatherViewModel @Inject constructor(
                         location.countryCode?.uppercase()
                     )
                 )
+
+                val appExpectation = result.exception.toAppException()
+                SnackbarManager.show(appExpectation.toMessageRes())
             }
 
             is WeatherResult.RefreshNotAvailable -> {
@@ -365,22 +428,24 @@ class WeatherViewModel @Inject constructor(
 
     private var autoRefreshJob: Job? = null
 
-    fun startAutoRefresh() {
+    fun startAutoRefresh(
+        location: Location,
+        source: Source
+    ) {
 
         if (autoRefreshJob?.isActive == true) return
 
         autoRefreshJob = viewModelScope.launch {
             while (isActive) {
 
-                delay(AUTO_REFRESH_INTERVAL)
-                val location = _uiState.value.activeLocation ?: continue
-
+                delay(45.minutes)
                 if (_uiState.value.isLoading || _uiState.value.isError) {
                     continue
                 }
 
                 getWeather(
-                    location = location
+                    location = location,
+                    source = source
                 )
             }
         }
